refactored arrow up to component, added to filtering section
