+++ frontend/src/app/modules/leader/leader-board/leader-board.component.html
@@ -16,8 +16,4 @@
         </div>
     </div>
 </div>
-<a pageScroll [pageScrollDuration]="1000" (pageScrollFinish)="afterScroll()" href="#top-scroll-page" *ngIf="showArrowButton">
-    <div class="scroll-top d-flex justify-content-center align-items-center">
-        <fa-icon [icon]="'arrow-up-long'" [size]="'lg'" class="scroll-top-icon"></fa-icon>
-    </div>
-</a>
+<app-arrow-up [onScrollEvent]="scrollEventSubject.asObservable()"></app-arrow-up>
