+++ frontend/src/app/modules/main/filtering-section/filtering-section.component.html
@@ -23,7 +23,7 @@
             (selectedItemsChanged)="onDifficultyChange($event)" class="m-0-12px"></app-dropdown-select>
         <app-dropdown-select [isMultiSelection]="true" [title]="'By tags'" [items]="tagsNames"
             (selectedItemsChanged)="onTagsChange($event)" class="m-0-12px mb-28px"></app-dropdown-select>
-        
+
     </div>
     <div class="challenges d-flex flex-column gap-2" infinite-scroll (scrolled)="onScroll()">
         <div class="d-flex text-center text-white text-opacity-50 mt-3" *ngIf="!challenges.length && !loading">
@@ -41,3 +41,4 @@
         </div>
     </div>
 </div>
+<app-arrow-up [onScrollEvent]="scrollEventSubject.asObservable()"></app-arrow-up>
