+++ frontend/src/app/modules/main/filtering-section/filtering-section.component.ts
@@ -16,7 +16,7 @@ import { ISortedModel } from '@shared/models/challenge/sorted-model';
 import { ILanguage } from '@shared/models/language/language';
 import { IPageSettings } from '@shared/models/page-settings';
 import { ITag } from '@shared/models/tag/tag';
-import { takeUntil } from 'rxjs';
+import { Subject, takeUntil } from 'rxjs';
 
 @Component({
     selector: 'app-filtering-section',
@@ -39,6 +39,8 @@ export class FilteringSectionComponent extends BaseComponent implements OnInit {
     public sortingLabels = SORTING_PROPERTIES.map((x) => x.label) as string[];
 
     public difficultyNames: string[];
+
+    public scrollEventSubject = new Subject<void>();
 
     private sortingProperty?: ISortedModel;
 
@@ -82,6 +84,8 @@ export class FilteringSectionComponent extends BaseComponent implements OnInit {
     }
 
     public onScroll() {
+        this.scrollEventSubject.next();
+
         if (this.isLastPage) {
             return;
         }
