+++ frontend/src/app/shared/components/arrow-up/arrow-up.component.html
@@ -1,0 +1,5 @@
+<a pageScroll [pageScrollDuration]="1000" (pageScrollFinish)="afterScroll()" href="#top-scroll-page" *ngIf="showArrowButton">
+    <div class="scroll-top d-flex justify-content-center align-items-center">
+        <fa-icon [icon]="'arrow-up-long'" [size]="'lg'" class="scroll-top-icon"></fa-icon>
+    </div>
+</a>
