+++ frontend/src/app/shared/components/arrow-up/arrow-up.component.sass
@@ -1,0 +1,13 @@
+@import 'color-variables'
+
+.scroll-top
+    background-color: $green
+    border-radius: 50%
+    position: fixed
+    right: 60px
+    bottom: 60px
+    width: 48px
+    height: 48px
+    cursor: pointer
+    &-icon
+        color: $white
