+++ frontend/src/app/shared/shared.module.ts
@@ -16,6 +16,7 @@ import { MonacoEditorModule } from 'ngx-monaco-editor-v2';
 import { NgxPageScrollModule } from 'ngx-page-scroll';
 import { NgxPageScrollCoreModule } from 'ngx-page-scroll-core';
 
+import { ArrowUpComponent } from './components/arrow-up/arrow-up.component';
 import { CodeEditorComponent } from './components/code-editor/code-editor.component';
 import { ConfirmationModalComponent } from './components/confirmation-modal/confirmation-modal.component';
 import { CustomInputComponent } from './components/custom-input/custom-input.component';
@@ -55,6 +56,7 @@ import { ProgressBarComponent } from './components/progress-bar/progress-bar.com
         ProgressBarComponent,
         NotificationsComponent,
         HeaderComponent,
+        ArrowUpComponent,
     ],
     exports: [
         CommonModule,
@@ -72,6 +74,7 @@ import { ProgressBarComponent } from './components/progress-bar/progress-bar.com
         CustomInputComponent,
         ProgressBarComponent,
         HeaderComponent,
+        ArrowUpComponent,
         NgbModule,
         TabsModule,
         MarkdownModule,
