+++ frontend/src/styles/global.sass
@@ -38,15 +38,3 @@ input:-webkit-autofill::first-line
 input:-webkit-autofill:focus
     -webkit-text-fill-color: $white
     caret-color: $white
-
-.scroll-top
-    background-color: $green
-    border-radius: 50%
-    position: fixed
-    right: 60px
-    bottom: 60px
-    width: 48px
-    height: 48px
-    cursor: pointer
-    &-icon
-        color: $white
